Fix exported env variables to ssh users (also include those with spaces)
New logic to pass env variables from the container to the SSH user. Now
instead of doing a fancy "grep" we do some even more fancy gathering which
variables are already set for the user and skip those explicitly.

We also skip multi-line variables because this is not supported by the
/etc/environment.